third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/touch-event-retargeting-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Touch event retargeting.

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
